backend/requirements.txt
scikit-learn==1.3.0
numpy==1.24.3
tensorflow==2.13.0  # якщо потрібні нейронні мережі
transformers==4.31.0  # для NLP
fastapi==0.104.1
sqlalchemy==2.0.23
"pydantic[email]==2.5.2"
pydantic-settings==2.1.0
httpx==0.27.0
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.11.3
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0
pytest==7.4.3
pytest-cov==4.1.0
